requirements.txt
alabaster==0.7.10
amqp==2.2.2
astroid==1.6.3
Babel==2.5.3
billiard==3.5.0.3
bleach==2.1.3
celery==4.1.0
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
dj-database-url==0.5.0
defusedxml==0.5.0
Django==1.9.13
django-appconf==1.0.2
django-bootstrap-breadcrumbs==0.9.1
django-bower==5.2.0
django-compressor==2.2
django-cors-headers==2.2.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-formtools==1.0
django-heroku==0.3.1
django-mobile==0.7.0
django-recaptcha==1.4.0
django-sortedm2m==1.5.0
django-tastypie==0.14.1
djangorestframework==3.2.5
docutils==0.14
easy-thumbnails==2.5
fitbit==0.3.0
gunicorn==19.8.1
django-recaptcha==1.3.1
django-tastypie==0.13.3
html5lib==1.0.1
icalendar==4.0.1
idna==2.6
imagesize==1.0.0
invoke==0.23.0
isort==4.3.4
Jinja2==2.10
kombu==4.1.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
oauthlib==2.0.7
packaging==17.1
pep8==1.7.1
Pillow==5.1.0
psycopg2==2.7.4
Pygments==2.2.0
pylint==1.8.4
pyparsing==2.2.0
python-dateutil==2.7.2
python-mimeparse==1.6.0
pytz==2018.4
rcssmin==1.0.6
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
rjsmin==1.0.12
simplejson==3.14.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
urllib3==1.22
vine==1.1.4
webencodings==0.5.1
whitenoise==3.3.1
wrapt==1.10.11
invoke==0.14.0
oauthlib==2.0.7
PyJWT==1.6.1
python3-openid==3.1.0
reportlab==3.3.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
## !! Could not determine repository location
wger==1.8
